+++ squirls-bootstrap/src/test/java/org/monarchinitiative/squirls/bootstrap/SquirlsFactoryTest.java
@@ -71,34 +71,60 @@
  *
  * version:6-8-18
  *
- * Daniel Danis, Peter N Robinson, 2020
+ * Daniel Danis, Peter N Robinson, 2021
  */
 
-package org.monarchinitiative.squirls.core.scoring.calculators.conservation;
+package org.monarchinitiative.squirls.bootstrap;
 
-import org.monarchinitiative.squirls.core.SquirlsException;
+import org.junit.jupiter.api.Test;
+import org.monarchinitiative.squirls.initialize.GenomicAssemblyVersion;
+import org.monarchinitiative.squirls.initialize.SquirlsResourceVersion;
+import org.monarchinitiative.squirls.io.SquirlsResourceException;
 
-/**
- * @author Daniel Danis
- */
-public class SquirlsWigException extends SquirlsException {
-    public SquirlsWigException() {
-        super();
+import java.io.File;
+
+import static org.hamcrest.Matchers.*;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+public class SquirlsFactoryTest {
+
+    private static final File RESOURCES_PATH = new File("src/test/resources/data");
+
+    private static final SimpleSquirlsProperties PROPERTIES = SimpleSquirlsProperties.builder(RESOURCES_PATH)
+            .annotatorProperties(new SimpleAnnotatorProperties())
+            .classifierProperties(new SimpleClassifierProperties())
+            .build();
+
+    @Test
+    public void constructor() throws Exception {
+        SquirlsConfigurationFactory factory = SquirlsConfigurationFactory.of(PROPERTIES);
+
+        assertThat(factory.supportedResourceVersions(), hasSize(1));
+        assertThat(factory.supportedResourceVersions(), hasItem(SquirlsResourceVersion.of("1710", GenomicAssemblyVersion.GRCH37)));
     }
 
-    public SquirlsWigException(String message) {
-        super(message);
+    @Test
+    public void getConfiguration() throws Exception {
+        SquirlsConfigurationFactory factory = SquirlsConfigurationFactory.of(PROPERTIES);
+        SquirlsResourceVersion resourceVersion = SquirlsResourceVersion.of("1710", GenomicAssemblyVersion.GRCH37);
+
+        Squirls configuration = factory.getSquirls(resourceVersion);
+
+        assertThat(configuration.resourceVersion(), is(equalTo(resourceVersion)));
+        assertThat(configuration.squirlsDataService(), is(notNullValue()));
+        assertThat(configuration.splicingAnnotator(), is(notNullValue()));
+        assertThat(configuration.squirlsClassifier(), is(notNullValue()));
+        assertThat(configuration.variantSplicingEvaluator(), is(notNullValue()));
     }
 
-    public SquirlsWigException(String message, Throwable cause) {
-        super(message, cause);
-    }
+    @Test
+    public void getConfiguration_throwsWhenMissing() throws Exception {
+        SquirlsConfigurationFactory factory = SquirlsConfigurationFactory.of(PROPERTIES);
+        SquirlsResourceVersion resourceVersion = SquirlsResourceVersion.of("1710", GenomicAssemblyVersion.GRCH38);
 
-    public SquirlsWigException(Throwable cause) {
-        super(cause);
-    }
-
-    public SquirlsWigException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
+        assertThat(factory.supportedResourceVersions(), not(hasItem(resourceVersion)));
+        SquirlsResourceException e = assertThrows(SquirlsResourceException.class, () -> factory.getSquirls(resourceVersion));
+        assertThat(e.getMessage(), matchesPattern("Resource 1710_hg38` is not present in `.*src/test/resources/data`"));
     }
 }