+++ squirls-initialize/src/main/java/org/monarchinitiative/squirls/initialize/MissingSquirlsResourceException.java
@@ -74,83 +74,34 @@
  * Daniel Danis, Peter N Robinson, 2020
  */
 
-package org.monarchinitiative.squirls.core.reference;
+package org.monarchinitiative.squirls.initialize;
 
-import org.monarchinitiative.svart.*;
-import org.monarchinitiative.svart.util.Seq;
-
-import java.util.Objects;
+import org.monarchinitiative.squirls.io.SquirlsResourceException;
 
 /**
+ * An exception thrown when a resource file (e.g. FASTA file) is missing from SQUIRLS data directory.
+ *
  * @author Daniel Danis
  */
-public class StrandedSequence extends BaseGenomicRegion<StrandedSequence> {
+public class MissingSquirlsResourceException extends SquirlsResourceException {
 
-    private final String sequence;
-
-    protected StrandedSequence(Contig contig, Strand strand, Coordinates coordinates, String sequence) {
-        super(contig, strand, coordinates);
-        this.sequence = sequence;
-        if (length() != sequence.length()) {
-            throw new IllegalArgumentException("Sequence length " + sequence.length() + " does not match length of the region " + length());
-        }
+    public MissingSquirlsResourceException() {
+        super();
     }
 
-    public static StrandedSequence of(GenomicRegion region, String sequence) {
-        return of(region.contig(), region.strand(), region.coordinates(), sequence);
+    public MissingSquirlsResourceException(String message) {
+        super(message);
     }
 
-    public static StrandedSequence of(Contig contig, Strand strand, Coordinates coordinates, String sequence) {
-        return new StrandedSequence(contig, strand, coordinates, sequence);
+    public MissingSquirlsResourceException(String message, Throwable cause) {
+        super(message, cause);
     }
 
-    public String sequence() {
-        return sequence;
+    public MissingSquirlsResourceException(Throwable cause) {
+        super(cause);
     }
 
-    /**
-     * @param query query region
-     * @return string with sequence that corresponds to <code>query</code> region or <code>null</code> if at least one
-     * base from the <code>region</code> is not available
-     */
-    public String subsequence(final GenomicRegion query) {
-        if (contains(query)) {
-            GenomicRegion queryOnStrand = query.withStrand(strand()).toZeroBased();
-            // when slicing a sequence, we always use 0-based coordinates - that's why we pre-calculate `start`
-            // field in the constructor
-            String seq = sequence.substring(
-                    queryOnStrand.start() - startWithCoordinateSystem(CoordinateSystem.zeroBased()),
-                    queryOnStrand.end() - startWithCoordinateSystem(CoordinateSystem.zeroBased()));
-            return query.strand() == strand()
-                    ? seq
-                    : Seq.reverseComplement(seq);
-        }
-        return null;
-    }
-
-    @Override
-    protected StrandedSequence newRegionInstance(Contig contig, Strand strand, Coordinates coordinates) {
-        return new StrandedSequence(contig, strand, coordinates, strand == strand() ? sequence : Seq.reverseComplement(sequence));
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        if (!super.equals(o)) return false;
-        StrandedSequence that = (StrandedSequence) o;
-        return Objects.equals(sequence, that.sequence);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(super.hashCode(), sequence);
-    }
-
-    @Override
-    public String toString() {
-        return "StrandedSequence{" +
-                "sequence='" + sequence + '\'' +
-                "} " + super.toString();
+    protected MissingSquirlsResourceException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
     }
 }