+++ squirls-initialize/src/main/java/org/monarchinitiative/squirls/initialize/SquirlsDataServiceImpl.java
@@ -71,86 +71,71 @@
  *
  * version:6-8-18
  *
- * Daniel Danis, Peter N Robinson, 2020
+ * Daniel Danis, Peter N Robinson, 2021
  */
 
-package org.monarchinitiative.squirls.core.reference;
+package org.monarchinitiative.squirls.initialize;
 
-import org.monarchinitiative.svart.*;
-import org.monarchinitiative.svart.util.Seq;
+import org.monarchinitiative.squirls.core.SquirlsDataService;
+import org.monarchinitiative.squirls.core.reference.StrandedSequence;
+import org.monarchinitiative.squirls.core.reference.StrandedSequenceService;
+import org.monarchinitiative.squirls.core.reference.TranscriptModel;
+import org.monarchinitiative.squirls.core.reference.TranscriptModelService;
+import org.monarchinitiative.svart.GenomicAssembly;
+import org.monarchinitiative.svart.GenomicRegion;
 
+import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 /**
  * @author Daniel Danis
  */
-public class StrandedSequence extends BaseGenomicRegion<StrandedSequence> {
+public class SquirlsDataServiceImpl implements SquirlsDataService {
 
-    private final String sequence;
+    private final StrandedSequenceService sequenceService;
+    private final TranscriptModelService transcriptModelService;
 
-    protected StrandedSequence(Contig contig, Strand strand, Coordinates coordinates, String sequence) {
-        super(contig, strand, coordinates);
-        this.sequence = sequence;
-        if (length() != sequence.length()) {
-            throw new IllegalArgumentException("Sequence length " + sequence.length() + " does not match length of the region " + length());
-        }
-    }
-
-    public static StrandedSequence of(GenomicRegion region, String sequence) {
-        return of(region.contig(), region.strand(), region.coordinates(), sequence);
-    }
-
-    public static StrandedSequence of(Contig contig, Strand strand, Coordinates coordinates, String sequence) {
-        return new StrandedSequence(contig, strand, coordinates, sequence);
-    }
-
-    public String sequence() {
-        return sequence;
-    }
-
-    /**
-     * @param query query region
-     * @return string with sequence that corresponds to <code>query</code> region or <code>null</code> if at least one
-     * base from the <code>region</code> is not available
-     */
-    public String subsequence(final GenomicRegion query) {
-        if (contains(query)) {
-            GenomicRegion queryOnStrand = query.withStrand(strand()).toZeroBased();
-            // when slicing a sequence, we always use 0-based coordinates - that's why we pre-calculate `start`
-            // field in the constructor
-            String seq = sequence.substring(
-                    queryOnStrand.start() - startWithCoordinateSystem(CoordinateSystem.zeroBased()),
-                    queryOnStrand.end() - startWithCoordinateSystem(CoordinateSystem.zeroBased()));
-            return query.strand() == strand()
-                    ? seq
-                    : Seq.reverseComplement(seq);
-        }
-        return null;
+    public SquirlsDataServiceImpl(StrandedSequenceService sequenceService, TranscriptModelService transcriptModelService) {
+        this.sequenceService = sequenceService;
+        this.transcriptModelService = transcriptModelService;
     }
 
     @Override
-    protected StrandedSequence newRegionInstance(Contig contig, Strand strand, Coordinates coordinates) {
-        return new StrandedSequence(contig, strand, coordinates, strand == strand() ? sequence : Seq.reverseComplement(sequence));
+    public GenomicAssembly genomicAssembly() {
+        return sequenceService.genomicAssembly();
+    }
+
+    @Override
+    public StrandedSequence sequenceForRegion(GenomicRegion region) {
+        return sequenceService.sequenceForRegion(region);
+    }
+
+    @Override
+    public List<String> getTranscriptAccessionIds() {
+        return transcriptModelService.getTranscriptAccessionIds();
+    }
+
+    @Override
+    public List<TranscriptModel> overlappingTranscripts(GenomicRegion query) {
+        return transcriptModelService.overlappingTranscripts(query);
+    }
+
+    @Override
+    public Optional<TranscriptModel> transcriptByAccession(String txAccession) {
+        return transcriptModelService.transcriptByAccession(txAccession);
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        if (!super.equals(o)) return false;
-        StrandedSequence that = (StrandedSequence) o;
-        return Objects.equals(sequence, that.sequence);
+        SquirlsDataServiceImpl that = (SquirlsDataServiceImpl) o;
+        return Objects.equals(sequenceService, that.sequenceService) && Objects.equals(transcriptModelService, that.transcriptModelService);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(super.hashCode(), sequence);
-    }
-
-    @Override
-    public String toString() {
-        return "StrandedSequence{" +
-                "sequence='" + sequence + '\'' +
-                "} " + super.toString();
+        return Objects.hash(sequenceService, transcriptModelService);
     }
 }