+++ squirls-spring-boot-autoconfigure/src/main/java/org/monarchinitiative/squirls/autoconfigure/AnnotatorPropertiesImpl.java
@@ -71,86 +71,32 @@
  *
  * version:6-8-18
  *
- * Daniel Danis, Peter N Robinson, 2020
+ * Daniel Danis, Peter N Robinson, 2021
  */
 
-package org.monarchinitiative.squirls.core.reference;
+package org.monarchinitiative.squirls.autoconfigure;
 
-import org.monarchinitiative.svart.*;
-import org.monarchinitiative.svart.util.Seq;
-
-import java.util.Objects;
+import org.monarchinitiative.squirls.core.scoring.SplicingAnnotator;
+import org.monarchinitiative.squirls.initialize.AnnotatorProperties;
+import org.springframework.boot.context.properties.ConfigurationProperties;
 
 /**
- * @author Daniel Danis
+ * Properties for tweaking the annotator.
  */
-public class StrandedSequence extends BaseGenomicRegion<StrandedSequence> {
-
-    private final String sequence;
-
-    protected StrandedSequence(Contig contig, Strand strand, Coordinates coordinates, String sequence) {
-        super(contig, strand, coordinates);
-        this.sequence = sequence;
-        if (length() != sequence.length()) {
-            throw new IllegalArgumentException("Sequence length " + sequence.length() + " does not match length of the region " + length());
-        }
-    }
-
-    public static StrandedSequence of(GenomicRegion region, String sequence) {
-        return of(region.contig(), region.strand(), region.coordinates(), sequence);
-    }
-
-    public static StrandedSequence of(Contig contig, Strand strand, Coordinates coordinates, String sequence) {
-        return new StrandedSequence(contig, strand, coordinates, sequence);
-    }
-
-    public String sequence() {
-        return sequence;
-    }
+@ConfigurationProperties(prefix = "squirls.annotator")
+public class AnnotatorPropertiesImpl implements AnnotatorProperties {
 
     /**
-     * @param query query region
-     * @return string with sequence that corresponds to <code>query</code> region or <code>null</code> if at least one
-     * base from the <code>region</code> is not available
+     * Which {@link SplicingAnnotator} to use (`agez` by default).
      */
-    public String subsequence(final GenomicRegion query) {
-        if (contains(query)) {
-            GenomicRegion queryOnStrand = query.withStrand(strand()).toZeroBased();
-            // when slicing a sequence, we always use 0-based coordinates - that's why we pre-calculate `start`
-            // field in the constructor
-            String seq = sequence.substring(
-                    queryOnStrand.start() - startWithCoordinateSystem(CoordinateSystem.zeroBased()),
-                    queryOnStrand.end() - startWithCoordinateSystem(CoordinateSystem.zeroBased()));
-            return query.strand() == strand()
-                    ? seq
-                    : Seq.reverseComplement(seq);
-        }
-        return null;
-    }
+    private String version = "agez";
 
     @Override
-    protected StrandedSequence newRegionInstance(Contig contig, Strand strand, Coordinates coordinates) {
-        return new StrandedSequence(contig, strand, coordinates, strand == strand() ? sequence : Seq.reverseComplement(sequence));
+    public String getVersion() {
+        return version;
     }
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        if (!super.equals(o)) return false;
-        StrandedSequence that = (StrandedSequence) o;
-        return Objects.equals(sequence, that.sequence);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(super.hashCode(), sequence);
-    }
-
-    @Override
-    public String toString() {
-        return "StrandedSequence{" +
-                "sequence='" + sequence + '\'' +
-                "} " + super.toString();
+    public void setVersion(String version) {
+        this.version = version;
     }
 }