+++ squirls-spring-boot-autoconfigure/src/main/java/org/monarchinitiative/squirls/autoconfigure/exception/MissingSquirlsResourceFailureAnalyzer.java
@@ -76,6 +76,7 @@
 
 package org.monarchinitiative.squirls.autoconfigure.exception;
 
+import org.monarchinitiative.squirls.initialize.MissingSquirlsResourceException;
 import org.springframework.boot.diagnostics.AbstractFailureAnalyzer;
 import org.springframework.boot.diagnostics.FailureAnalysis;